Changes made years ago to make it work on my raspberry pi 2

+++ client/stt.py
@@ -15,6 +15,7 @@ import yaml
 import jasperpath
 import diagnose
 import vocabcompiler
+import pdb
 
 
 class AbstractSTTEngine(object):
@@ -122,8 +123,14 @@ class PocketSphinxSTT(AbstractSTTEngine):
                                  "hmm_dir in your profile.",
                                  hmm_dir, ', '.join(missing_hmm_files))
 
-        self._decoder = ps.Decoder(hmm=hmm_dir, logfn=self._logfile,
-                                   **vocabulary.decoder_kwargs)
+	config = ps.Decoder.default_config()
+	config.set_string('-hmm', hmm_dir)
+	config.set_string('-logfn', self._logfile)
+	config.set_string('-lm', vocabulary.decoder_kwargs['lm'])
+	config.set_string('-dict', vocabulary.decoder_kwargs['dict'])
+        self._decoder = ps.Decoder(config)
+        #self._decoder = ps.Decoder(hmm=hmm_dir, logfn=self._logfile,
+                                   #**vocabulary.decoder_kwargs)
 
     def __del__(self):
         os.remove(self._logfile)
@@ -163,13 +170,15 @@ class PocketSphinxSTT(AbstractSTTEngine):
         self._decoder.process_raw(data, False, True)
         self._decoder.end_utt()
 
-        result = self._decoder.get_hyp()
+        result = self._decoder.hyp()
+	if result:
+	    result = result.hypstr
         with open(self._logfile, 'r+') as f:
             for line in f:
                 self._logger.debug(line.strip())
             f.truncate()
 
-        transcribed = [result[0]]
+        transcribed = [result]
         self._logger.info('Transcribed: %r', transcribed)
         return transcribed
